Make more robust
There was wisdom in the go-qbittorrent method
ReannounceTorrentWithRetry.  First, when the torrent
is fresh (3s), the t.Url is empty.  Second, the Torrent
struct is not useful for this purpose, because its NumSeeds is always 0.
So just pass around the hash and use GetTorrentTrackersCtx to get NumSeeds.
Here is where this impl differs:
1. it runs outside the autobrr hook, so the hook is free to move
   on and write the release status
2. it iterates until the tracker is ok AND there are >0 seeds
3. it iterates 2 extra times like the deluge-reannounce script