Add more logging to reannounce loops
It is now my belief that once one tracker is OK, then
calling ReAnnounceTorrents is a noop.  I see that
libtorrent keeps a min_announce in addition to next_announce,
but only the qBittorrent thick client displays it, and it
seems not to change after calling ReAnnounceTorrents.
So this was an exercise in learning.